change compare test to not use json encoding and decoding
